Move schema to separate file
